refactor(low-code): move authenticator capability protocols next to the authenticators
`ResponseAwareAuthenticator` and `TokenRotatingAuthenticator` described what an
authenticator can offer, but lived in `http_client.py` because that is what calls
them. Anyone writing an authenticator would look under `requests_native_auth/`,
which is where the base classes live and where `DeclarativeAuthenticator` inherits
from, so that is where the contracts belong.

Moved to `requests_native_auth/protocols.py` and exported from the package. This
costs nothing in the import graph: `HttpClient` dispatches structurally through
`getattr`, so it never referenced the protocol objects at runtime -- only a docstring
mentioned them. `streams/http` still imports nothing from `declarative`, and the
protocols' new home does not import `HttpClient`, so an authenticator that needs the
client (as the declarative one does) cannot create a cycle.

`RateLimitedMultipleTokenAuthenticator` now names the two protocols it implements, so
the contract is reachable from either end.
